adds unittest for node metrics

+++ metricbeat/module/couchbase/node/data.go
@@ -8,28 +8,28 @@ import (
 )
 
 type NodeSystemStats struct {
-	CPUUtilizationRate int `json:"cpu_utilization_rate"`
+	CPUUtilizationRate float32 `json:"cpu_utilization_rate"`
 	SwapTotal          int64 `json:"swap_total"`
-	SwapUsed           int `json:"swap_used"`
+	SwapUsed           int64 `json:"swap_used"`
 	MemTotal           int64 `json:"mem_total"`
 	MemFree            int64 `json:"mem_free"`
 }
 
 type NodeInterestingStats struct {
-	CmdGet                   int `json:"cmd_get"`
-	CouchDocsActualDiskSize  int `json:"couch_docs_actual_disk_size"`
-	CouchDocsDataSize        int `json:"couch_docs_data_size"`
-	CouchSpatialDataSize     int `json:"couch_spatial_data_size"`
-	CouchSpatialDiskSize     int `json:"couch_spatial_disk_size"`
-	CouchViewsActualDiskSize int `json:"couch_views_actual_disk_size"`
-	CouchViewsDataSize       int `json:"couch_views_data_size"`
-	CurrItems                int `json:"curr_items"`
-	CurrItemsTot             int `json:"curr_items_tot"`
-	EpBgFetched              int `json:"ep_bg_fetched"`
-	GetHits                  int `json:"get_hits"`
-	MemUsed                  int `json:"mem_used"`
-	Ops                      int `json:"ops"`
-	VbReplicaCurrItems       int `json:"vb_replica_curr_items"`
+	CmdGet                   int64 `json:"cmd_get"`
+	CouchDocsActualDiskSize  int64 `json:"couch_docs_actual_disk_size"`
+	CouchDocsDataSize        int64 `json:"couch_docs_data_size"`
+	CouchSpatialDataSize     int64 `json:"couch_spatial_data_size"`
+	CouchSpatialDiskSize     int64 `json:"couch_spatial_disk_size"`
+	CouchViewsActualDiskSize int64 `json:"couch_views_actual_disk_size"`
+	CouchViewsDataSize       int64 `json:"couch_views_data_size"`
+	CurrItems                int64 `json:"curr_items"`
+	CurrItemsTot             int64 `json:"curr_items_tot"`
+	EpBgFetched              int64 `json:"ep_bg_fetched"`
+	GetHits                  int64 `json:"get_hits"`
+	MemUsed                  int64 `json:"mem_used"`
+	Ops                      int64 `json:"ops"`
+	VbReplicaCurrItems       int64 `json:"vb_replica_curr_items"`
 }
 
 type Node struct {
@@ -38,14 +38,14 @@ type Node struct {
 	Uptime             string `json:"uptime"`
 	MemoryTotal        int64 `json:"memoryTotal"`
 	MemoryFree         int64 `json:"memoryFree"`
-	McdMemoryReserved  int `json:"mcdMemoryReserved"`
-	McdMemoryAllocated int `json:"mcdMemoryAllocated"`
+	McdMemoryReserved  int64 `json:"mcdMemoryReserved"`
+	McdMemoryAllocated int64 `json:"mcdMemoryAllocated"`
 	ClusterMembership  string `json:"clusterMembership"`
 	RecoveryType       string `json:"recoveryType"`
 	Status             string `json:"status"`
 	ThisNode           bool `json:"thisNode"`
 	Hostname           string `json:"hostname"`
-	ClusterCompatibility int `json:"clusterCompatibility"`
+	ClusterCompatibility int64 `json:"clusterCompatibility"`
 	Version              string `json:"version"`
 	Os                   string `json:"os"`
 }
